pyproject.toml
[project]
name = "ForSureLLM"
version = "0.1.0"
description = "YES/NO/unknown classifier"
requires-python = ">=3.10"
dependencies = [
    "onnxruntime>=1.17",
    "tokenizers>=0.15",
    "numpy>=1.24",
]

[project.optional-dependencies]
train = [
    "litellm>=1.50",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "transformers>=4.40",
    "datasets>=2.18",
    "torch>=2.2",
    "optimum[onnxruntime]>=1.19",
    "scikit-learn>=1.4",
]
web = [
    "fastapi>=0.100",
    "uvicorn[standard]>=0.20",
]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["forsurellm"]
